requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Pillow==10.0.1
numpy==1.24.3
redis==4.5.4
flask-limiter==3.5.0
prometheus-client==0.17.1
flask-cors==4.0.0 